- navbar refactors - the tests need to be revisited and completed
